Split and refactored IQuestLog, QuestController and tests in preparation for database setup
