+++ HikingQuests.Server/Controllers/QuestWorkflowController.cs
@@ -1,0 +1,30 @@
+﻿using HikingQuests.Server.Infrastructure.Entities;
+using HikingQuests.Server.Models;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+
+namespace HikingQuests.Server.Controllers
+{
+    [Route("api/quest-workflow")]
+    [ApiController]
+    public class QuestWorkflowController : ControllerBase
+    {
+        private readonly IQuestWorkflowService _questWorkFlowService;
+        public QuestWorkflowController(IQuestWorkflowService questWorkFlowService) 
+            => _questWorkFlowService = questWorkFlowService;
+
+        [HttpPatch("{id}/start")]
+        public IActionResult StartQuest(Guid id)
+        {
+            var updatedQuest = _questWorkFlowService.StartQuest(id);
+            return Ok(updatedQuest);
+        }
+
+        [HttpPatch("{id}/complete")]
+        public IActionResult CompleteQuest(Guid id)
+        {
+            _questWorkFlowService.CompleteQuest(id);
+            return NoContent();
+        }
+    }
+}
