+++ HikingQuests.Server/Models/IQuestQueryService.cs
@@ -1,0 +1,9 @@
+﻿namespace HikingQuests.Server.Models
+{
+    public interface IQuestQueryService
+    {
+        QuestItem GetQuestById(Guid id);
+        QuestItem GetQuestByTitle(string title);
+        IEnumerable<QuestItem> GetAllQuestItems();
+    }
+}
