+++ HikingQuests.Server/Models/IQuestWorkflowService.cs
@@ -1,0 +1,8 @@
+﻿namespace HikingQuests.Server.Models
+{
+    public interface IQuestWorkflowService
+    {        
+        QuestItem StartQuest(Guid id);
+        void CompleteQuest(Guid id);
+    }
+}
